20190917 - html 코드를 naverCafeWrtier로 이관
